requirements.txt
slackclient
google-cloud-translate